requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
starlette==0.27.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
email-validator==2.1.0
python-multipart==0.0.24